% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BehavioralDataAnalysis-package.R
\docType{package}
\name{BehavioralDataAnalysis-package}
\alias{BehavioralDataAnalysis}
\alias{BehavioralDataAnalysis-package}
\title{BehavioralDataAnalysis: Bootstrap and Sampling Functions For Behavioral Data Analysis}
\description{
Based on the book Behavioral Data Analysis With R and Python. It provides robust functions to analyze behavioral data without relying on traditional statistics.
}
\author{
\strong{Maintainer}: Florent Buisson \email{florent.buisson.oreilly@maskedmails.com}

}
\keyword{internal}
