% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BehavioralDataAnalysis-package.R
\docType{package}
\name{BehavioralDataAnalysis}
\alias{BehavioralDataAnalysis}
\title{BehavioralDataAnalysis}
\description{
Description of your package
}
\author{
Florent <florent.buisson.oreilly@maskedmails.com>
}
