% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_CI.R
\name{boot_CI}
\alias{boot_CI}
\title{Creates a Bootstrap confidence interval for the output of a function}
\usage{
boot_CI(df, fct, B = 100, conf.level = 0.9)
}
\arguments{
\item{df}{data to use}

\item{fct}{function to run on the data}

\item{B}{number of Bootstrap sampling iterations to run}

\item{conf.level}{confidence interval percentage to use}
}
\value{
a vector with two values, the lower and upper bounds of the confidence interval
}
\description{
Creates a Bootstrap confidence interval for the output of a function
}
\examples{
df1 <- data.frame(
id = 1:6,
x = c(1, 1.5, 5, 5.5, 10, 10.5),
y = c(1, 1.5, 5, 5.5, 10, 10.5))
boot_CI(df1, function(df) mean(df$x), B=100)
}
