% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_char}
\alias{df_char}
\title{Data frame with character variable}
\format{
## A data frame with 6 observations of 4 variables:
\describe{
 \item{id}{unique row identifier}
 \item{x}{numeric variable}
 \item{y}{numeric variable}
 \item{z}{character variable}
}
A small data frame with numeric and character variables
}
\usage{
df_char
}
\description{
Data frame with character variable
}
\keyword{datasets}
