% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_na}
\alias{df_na}
\title{A small data frame with only numeric variables and some NA values}
\format{
## A data frame with 6 observations of 3 variables:
\describe{
 \item{id}{unique row identifier}
 \item{x}{numeric variable}
 \item{y}{numeric variable}
}
}
\usage{
df_na
}
\description{
A small data frame with only numeric variables and some NA values
}
\keyword{datasets}
