% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_num}
\alias{df_num}
\title{Data frame with numeric variables}
\format{
## A data frame with 6 observations of 3 variables:
\describe{
 \item{id}{unique row identifier}
 \item{x}{numeric variable}
 \item{y}{numeric variable}
}
A small data frame with only numeric variables
}
\usage{
df_num
}
\description{
Data frame with numeric variables
}
\keyword{datasets}
