% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.R
\name{pairing}
\alias{pairing}
\title{Create a stratified assignment (aka. pair matching)}
\usage{
pairing(df, id, n.groups = 2)
}
\arguments{
\item{df}{data to use}

\item{id}{column(s) used for identification}

\item{n.groups}{number of groups}
}
\value{
a list of pairs, each having n.groups elements
}
\description{
Create a stratified assignment (aka. pair matching)
}
\examples{
df1 <- data.frame(
id = 1:6,
x = c(1, 1.5, 5, 5.5, 10, 10.5),
y = c(1, 1.5, 5, 5.5, 10, 10.5))
pairing(df1, 'id')
}
