% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strat_assign.R
\name{strat_assign}
\alias{strat_assign}
\title{Create a stratified assignment (aka. pair matching)}
\usage{
strat_assign(df, id, n.groups = 2)
}
\arguments{
\item{df}{data to use}

\item{id}{column(s) used for identification}

\item{n.groups}{number of groups}
}
\value{
the original data frame, with an added column for group assignment
}
\description{
Create a stratified assignment (aka. pair matching)
}
\examples{
df1 <- data.frame(
id = 1:6,
x = c(1, 1.5, 5, 5.5, 10, 10.5),
y = c(1, 1.5, 5, 5.5, 10, 10.5))
strat_assign(df1, 'id')
}
